BINIUS-321: group BaseFold and the IOP channels into module folders
The iop and iop-prover crates already keep FRI in a tidy fri/ module
folder, but BaseFold sat as three loose flat files and the channel
abstraction was scattered: the IOP*Channel trait next to its unrelated
test/estimation impls. The directory listing hid the structure.

Mirror the fri/ pattern:

- basefold/ = mod.rs (facade) + opening.rs + channel.rs + compiler.rs,
  plus error.rs on the verifier side (mirrors fri/error.rs).
- channel/ = mod.rs (the IOP*Channel trait) + naive.rs, and
  size_tracking.rs + oracle_setup.rs on the verifier side.

merkle_channel stays its own module: it is a distinct transport trait
that both FRI and BaseFold sit on, not a peer impl of IOP*Channel.

Pure move + rename, no logic change:
- files moved with git mv (history preserved as renames);
- inline #[cfg(test)] mod tests travel with their files, no tests.rs;
- no compatibility aliases -- every call site renamed to the new paths
  across spartan-*, prover, verifier, m4-*, and the internal crates;
- verbose //! header docs trimmed to one-liners.
