+++ crates/iop-prover/src/basefold/mod.rs
@@ -1,0 +1,10 @@
+// Copyright 2025 Irreducible Inc.
+// Copyright 2026 The Binius Developers
+
+//! Prover for the BaseFold sumcheck-PIOP to IP compiler.
+
+pub mod channel;
+pub mod compiler;
+mod opening;
+
+pub use opening::*;
