+++ crates/iop-prover/src/lib.rs
@@ -26,11 +26,8 @@
 #![warn(rustdoc::missing_crate_level_docs)]
 
 pub mod basefold;
-pub mod basefold_channel;
-pub mod basefold_compiler;
 pub mod channel;
 pub mod fri;
 pub mod logup_star;
 pub mod merkle_channel;
 pub mod merkle_tree;
-pub mod naive_channel;
