+++ crates/iop/src/basefold/mod.rs
@@ -1,0 +1,29 @@
+// Copyright 2025 Irreducible Inc.
+// Copyright 2026 The Binius Developers
+
+//! Verifier for the BaseFold sumcheck-PIOP to IP compiler.
+//!
+//! [BaseFold] is a generalized polynomial commitment scheme that allows compilation of
+//! sumcheck-PIOP protocols to IOPs. The protocol is an interactive argument for sumcheck claims
+//! of multivariate polynomials defined as the product of a committed multilinear polynomial and a
+//! transparent multilinear polynomial. When the transparent polynomial is a multilinear equality
+//! indicator, this BaseFold instance becomes a multilinear polynomial commitment scheme. The core
+//! idea is to commit the multilinear polynomial using FRI and open the sumcheck claim using an
+//! interleaved instance of sumcheck on the composite polynomial and FRI on the committed codeword,
+//! sharing folding challenges.
+//!
+//! This module implements the version specialized for binary field FRI described in [DP24],
+//! Section 4. Moreover, this module includes the classic [BCS16] compiler for IOPs to IPs that
+//! commits and opens oracle messages using Merkle trees.
+//!
+//! [BaseFold]: <https://link.springer.com/chapter/10.1007/978-3-031-68403-6_5>
+//! [DP24]: <https://eprint.iacr.org/2024/504>
+//! [BCS16]: <https://eprint.iacr.org/2016/116>
+
+pub mod channel;
+pub mod compiler;
+mod error;
+mod opening;
+
+pub use error::*;
+pub use opening::*;
